Hook to simplify MOOC creation
In order to simplify the process of adding a MOOC, this change adds a
hook for page creation within the MOOC namespace.
Whenever a root MOOC page is created, a MoocOverview stub is injected
into the edit box.
Together with this change, the MoocEntity class tree has slighlty be
changed to allow a convenient usage of these classes inside the hook.
Additionally, the structure loading query has been fixed together with
the stub of lessons in the edit-JS.

Change-Id: I28538a52af765d0fe88dfba2dc2a565ffbd873f6