UI improvements
In this change the interface and its usability is improved by
making the section content expanders working again, especially
using CSS3. Units in lessons and overviews are now shown with the
same height, regardless if having a video or not. In addition,
their preview boxes are now clickable and refer to the unit.
Video players are replaced with a linked thumbnail of the video.
If the video is missing, a suitable alternative text is shown.
Messages are now preloaded by adding them to the module.

Change-Id: I34d295773e12800646fee9c077d9487174fbcc8a